Implemented allowed congestion

+++ scheduler.go
@@ -292,7 +292,7 @@ func (sch *scheduler) selectPathDQNAgent(s *session, hasRetransmission bool, has
 
 
 	for pathID, pth := range s.paths{
-		cong := pth.sentPacketHandler.GetCongestionWindow()-pth.sentPacketHandler.GetBytesInFlight()
+		cong := float32(pth.sentPacketHandler.GetCongestionWindow())-float32(pth.sentPacketHandler.GetBytesInFlight())
 		allowed := pth.SendingAllowed() || (cong <= 0 && float32(cong) >=  -float32(pth.sentPacketHandler.GetCongestionWindow()) * float32(sch.AllowedCongestion) * 0.01)
 		if !pth.SendingAllowed(){
 			utils.Errorf("Not allowed: %f, %f", float32(cong), -float32(pth.sentPacketHandler.GetCongestionWindow()) * float32(sch.AllowedCongestion) * 0.01)
